Changed subsetting by position to name
The previous version assumed that the items element is in the fifth position. However, the YouTube API seems to return items in the fourth position now breaking the code. This patch changes from position to name-based subsetting. I also simplified the code and removed the dependency to plyr by using base functions.