Only check video parameters if type is "video" (#120)
For video_license, video_syndicated, and video_type, only enforce the
guard clauses if the type parameter is set to "video".  The values in
these parameters are only used when `type == "video"`, otherwise the
variables are coerced to NULL before being sent to the API.

The check for `type == "video"` is required in order to prevent errors in
the guard clauses when paginating on requests with type not equal to
"video".  Since the video parameters are set to NULL when type is not
"video", if pagination is required then then call to `yt_search` for the
second page will have the video_* parameters set to NULL and cause an
error.

Co-authored-by: Juan Alvarado <wok1@cdc.gov>